requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
minio
python-dotenv
